+++ src/platform/w3dfilesystem.h
@@ -68,8 +68,8 @@ public:
     virtual bool Open(const char *filename, int rights = FM_READ);
     virtual bool Open(int rights = FM_READ);
     virtual int Read(void *buffer, int length);
-    virtual int Seek(int offset, int whence = FS_SEEK_CURRENT);
-    virtual int Size();
+    virtual off_t Seek(off_t offset, int whence = FS_SEEK_CURRENT);
+    virtual off_t Size();
     virtual int Write(void const *buffer, int size);
     virtual void Close();
